Update SafeThanSorry.py
here is the tentative code for safe than sorry. I updated this file to indule the quizzing section of the code. This should be able to capture the users name and allow them to pick whichever module of the 4 they want to be quizzed on. in the end, the code will give their results in a completion message.