Update SafeThanSorry.py
When I tested the program, I found many issues with it running.  This is the updated code. I found that the new lines that I used the enter key for actually needed to be put into their own line quotes in order for it to run. i corrected each line of the code that lists the questions and answers. i then corrected the spacing of the output putting gaps between questions and answer choices. Currently I am trying to get the output to display a percentage score.