Update SafeThanSorry.py
This update includes major and minor changes. I did some wording and formatting fixes. 

This was updated so that the user has more autonomy over the order in which they can complete the quiz modules. this is simply for personalization.

I figured out how to report the score. i had to define the score within the function in order for the score calculation to be correct and display. I also had to create a separate function to have a final report and call it. 

i also have current progress displayed after each question under the testing function. and report final score at the end.

This code words as long as the user follows instructions. i ran into an issue because I cant get output how I want it if the user enters empty string, out of range number module, and any letters. it wont call the function again and loop until the right information is entered how I want it to.